Add Project 3
This was a deliberately short technique project taking an existing app and making it better. I hope you didn't get too bored, and hope even more that some of the new material sunk in because we covered UIBarButtonItem and UIActivityViewController.I hope you can see how trivial it is to add social media to your apps, and it can make a huge difference to helping spread the word about your work once your apps are on the App Store. I hope this project has also shown you how easy it is to go back to previous projects and improve them with only a little extra effort.